defBrowser: Base on Max's implementation to do rewriting
Rename and some code rewrite base on Max's implementation.
